backend/requirements.txt
celery==5.2.2
asgiref==3.4.1
bio==1.3.3
biopython==1.79
biothings-client==0.2.6
certifi==2021.10.8
charset-normalizer==2.0.9
Django==4.0
django-cors-headers==3.10.1
djangorestframework==3.13.1
idna==3.3
install==1.3.5
mygene==3.2.2
numpy==1.21.5
pytz==2021.3
requests==2.26.0
redis==4.1.0
sqlparse==0.4.2
tqdm==4.62.3
urllib3==1.26.7